move text-render out of app.cpp
